backend/blueprints/ml_data_pipeline/examples/cats-batch_example/resources/img_manipulation/requirements.txt
Pillow
boto3
